# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "termsOfService": "https://github.com/Team-innovation-INC/PregressServer.git",
    "version": "1.0.0",
    "title": "progress_app",
    "description": "node server for progress app application",
    "contact": {
      "name": "Raed Rdhaounia",
      "email": "raedrdhaounia@gmail.com",
      "url": "https://raedrdhaounia.netlify.app/"
    },
    "license": {
      "name": "ISC",
      "url": "https://github.com/Team-innovation-INC/demo_Progress/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "http://{host}:{port}/",
      "description": "local server for development",
      "variables": {
        "host": {
          "default": "localhost",
          "description": "this value is assigned by the service provider, in this example `localhost:5000`"
        },
        "port": {
          "enum": [
            5000,
            3000
          ],
          "default": 5000
        }
      }
    },
    {
      "url": "https://progress-application.onrender.com/",
      "description": "deployed server for the app "
    },
    {
      "url": "https://progress-develop-server.onrender.com/",
      "description": "developer server for test"
    }
  ],
  "basePath": "",
  "paths": {
...
    "/api/company/test": {
      "get": {
        "tags": [
          "-- AUTHENTICATION COMPANY --"
        ],
        "summary": "test router for the the test path",
        "description": "this route is for testing the auth tag routes have no use for the app just for testing.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/company/create": {
      "post": {
        "tags": [
          "-- AUTHENTICATION COMPANY --"
        ],
        "summary": "summary for the company route...",
        "description": "description for the company route...",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "companyName": {
                    "example": "any"
                  },
                  "companyWebSite": {
                    "example": "any"
                  },
                  "bio": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/company/activate": {
      "get": {
        "tags": [
          "-- AUTHENTICATION COMPANY --"
        ],
        "summary": "summary for the company route...",
        "description": "description for the company route...",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "referer",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/company/join": {
      "post": {
        "tags": [
          "-- AUTHENTICATION COMPANY --"
        ],
        "summary": "summary for the company route...",
        "description": "description for the company route...",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "companyId": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/company/list": {
      "get": {
        "tags": [
          "-- AUTHENTICATION COMPANY --"
        ],
        "summary": "summary for the company route...",
        "description": "description for the company route...",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/company/delete": {
      "delete": {
        "tags": [
          "-- AUTHENTICATION COMPANY --"
        ],
        "summary": "summary for the company route...",
        "description": "description for the company route...",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/company/invite": {
      "post": {
        "tags": [
          "-- AUTHENTICATION COMPANY --"
        ],
        "summary": "summary for the company route...",
        "description": "description for the company route...",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
  "security": [
    {
      "bearerAuth": []
    }
  ]
}